requirements.txt
Flask==2.2.3
python-dateutil==2.8.2
gunicorn==20.1.0